chore: workflow and model-service cleanup (#1313)
