--- packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/models/modelservice/Graph.java
@@ -1,14 +1,0 @@
-package software.uncharted.terarium.hmiserver.models.modelservice;
-
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-import java.io.Serializable;
-import java.util.List;
-
-@Data
-@Accessors(chain = true)
-public class Graph implements Serializable {
-	private List<Node> nodes;
-	private List<Edge> edges;
-}
