--- packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/models/modelservice/Node.java
@@ -1,13 +1,0 @@
-package software.uncharted.terarium.hmiserver.models.modelservice;
-
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-import java.io.Serializable;
-
-@Data
-@Accessors(chain = true)
-public class Node implements Serializable {
-	private String name;
-	private String type;
-}
