--- packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/models/modelservice/SimulateParams.java
@@ -1,14 +1,0 @@
-package software.uncharted.terarium.hmiserver.models.modelservice;
-
-import lombok.Data;
-import lombok.experimental.Accessors;
-
-import java.io.Serializable;
-import java.util.Map;
-
-@Data
-@Accessors(chain = true)
-public class SimulateParams implements Serializable {
-	private Map<String, Double> variables;
-	private Map<String, Double> parameters;
-}
